move enums to correct file
