+++ rdr_service/model/study_nph.py
@@ -4,9 +4,9 @@ from sqlalchemy import Column, Integer, BigInteger, String, ForeignKey, Index, e
 from sqlalchemy.dialects.mysql import TINYINT, JSON
 from sqlalchemy.orm import relation, relationship
 
-from rdr_service.ancillary_study_resources.nph.enums import ConsentOptInTypes, ParticipantOpsElementTypes
+from rdr_service.ancillary_study_resources.nph.enums import ConsentOptInTypes, ParticipantOpsElementTypes, \
+    StoredSampleStatus, IncidentStatus, IncidentType
 from rdr_service.model.base import NphBase, model_insert_listener, model_update_listener
-from rdr_service.model.study_nph_enums import StoredSampleStatus, IncidentStatus, IncidentType
 from rdr_service.model.utils import UTCDateTime, Enum
 
 
@@ -313,6 +313,24 @@ class DeactivationEvent(NphBase):
 
 event.listen(DeactivationEvent, "before_insert", model_insert_listener)
 event.listen(DeactivationEvent, "before_update", model_update_listener)
+
+
+# class DietEvent(NphBase):
+#     __tablename__ = "deactivation_event"
+#
+#     id = Column("id", BigInteger, autoincrement=True, primary_key=True)
+#     created = Column(UTCDateTime)
+#     modified = Column(UTCDateTime)
+#     ignore_flag = Column(TINYINT, default=0)
+#     event_authored_time = Column(UTCDateTime)
+#     participant_id = Column(BigInteger, ForeignKey("participant.id"))
+#     event_id = Column(BigInteger, ForeignKey("participant_event_activity.id"))
+#     external_id = Column(String(32))
+#     diet_name
+#
+#
+# event.listen(DeactivationEvent, "before_insert", model_insert_listener)
+# event.listen(DeactivationEvent, "before_update", model_update_listener)
 
 
 class SampleUpdate(NphBase):
